putting cards, faces,icons in memory and initializing them at the start of game.
Commented out old icons tex reference.

per wiki i added font color id enum to TIM for the menu fonts

changed some things to ushort as it can't be larger or smaller than one. I read that fonts had a max clutcount of 8. I wonder what the if the others have similar restictions

if a texture with the same name exists in the textures folder like from a mod you can use texturehandler to load it. Now I need to add support for multipart textures.

I ment to commit a while ago but i kept going and lost track of time.

oh made scale a vector2[] as not all textures will have upressed versions of them. so I need to track the scale change for each texture I upscale. atleast for atlas's some other textures we decide the output dims so we don't care what the scale is. as long as the aspec ratio is correct.

since we don't always need the scale i added overloads with out it.

+++ FF8/module_battle_Debug.cs
@@ -642,7 +642,7 @@ namespace FF8
         {
             textureInterface = new TIM2(stageBuffer, texturePointer);
             textures = new Texture2D[textureInterface.GetClutCount];
-            for (int i = 0; i < textureInterface.GetClutCount; i++)
+            for (ushort i = 0; i < textureInterface.GetClutCount; i++)
             {
                 byte[] b = textureInterface.CreateImageBuffer(textureInterface.GetClutColors(i));
                 Texture2D tex = new Texture2D(Memory.spriteBatch.GraphicsDevice,
